backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cryptography
